[Crashtracking] Collect PDB info (#6176)
## Summary of changes

On Windows, when generating a crash report, collect the PDB info
(signature, age).

## Reason for change

Needed to resolve the symbols on server side.

## Implementation details

In theory, we could have used `SymGetModuleInfoW64` from dbghelp to
retrieve that information, but the returned `IMAGEHLP_MODULEW64`
structure has evolved over time so we can't rely on the version of the
DLL bundled with Windows. We could ship the dbghelp from the Windows SDK
but that's a lot of extra space for an information that we can retrieve
by other means.

So instead, this PR parses the PE header of the loaded modules to
retrieve the PDB info. One source of complexity is that we're processing
a remote process, so we have to use `ReadProcessMemory` first to fetch
the memory we need. I introduced an abstraction to be able to test it
locally in unit tests.

## Test coverage

- Added unit tests to validate that we correctly parse a PE32 and a PE64
image (I had to embed DLLs into the test project for that)
- Added assertions to the integration tests to make sure we properly
extract the PDB information