Fixes Clion build on OSX (#6215)
## Summary of changes

This PR fixes an issue when using CLion on OSX

## Reason for change

the CMakeList.txt import the glibc-compat project causing the issue in
OSX due the change on the C version.

## Implementation details

We skip the definition of this project for osx because is not required
in the operating system.

## Test coverage

If the pipeline is green then everything works correctly.

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->