[ASM] Add event rules version in telemetry (#6208)
## Summary of changes

In the WAF related telemetry metrics, we are not currently sending the
rules version, only the WAF version. All the other libraries already
send that and we have been requested to send this metric as well.

For now, the tag is only sent in non-RASP ASM metrics. RASP metrics will
include this tag in the near future but the it is still not defined.

## Reason for change

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->