[DBM][fix] Do not prepend DBM injected comment when a pg plan hint is present (#6204)
## Summary of changes

Postgres supports hints for execution plans, provided they are in the
*first comment*, and start with `/*+`.
If we prepend the DBM injected comment in those cases, the hint is not
the first comment anymore, and is not taken into account.
This PR fixes that problem by detecting the hints when using Npgsql, and
when we do, we append the comment instead of prepending it.

## Implementation details

Took the opportunity to integrate modifying the comment into
`DatabaseMonitoringPropagator`, so that it's all self-contained. This
requires passing the `IDbCommand` as a parameter so that we can set the
property.
I made sure to also edit the check on reused commands to also detect
appended comments (with some extra conditions to avoid "paying the
price" of a full string search every time).

I recommend reviewing with "ignore whitespace changes" active, as I
flipped the `if` condition in `PropagateDataViaComment` to avoid having
the whole body of the method indented by an extra level.

## Test coverage

yes.