Update version conflict test (#6226)
## Summary of changes

Change the version conflict test to something sane

## Reason for change

Since 3.x, this test doesn't really make sense - it's meant to test the
behaviour of the app when the _manual_ version is _higher_ than the
automatic version. Since 3.x, we don't ship Datadog.Trace.dll in the
NuGet package, so this scenario, where the manual version is _higher_
than the automatic, can't happen in practice.

What we _can_ still have is a manual version 2.x with automatic 3.x, so
changed the test to check that scenario instead.

## Implementation details

- Stop using the magic "future" NuGet and use the existing latest 2.x
line.
- The behaviour should be the same for both .NET FX and .NET Core now

## Test coverage

Kinda the same, kinda different. We're testing auto > manual instead of
manual > auto now.

## Other details

This is particularly necessarily for supporting .NET 9, seeing as the
existing 255.x.x package won't support it, so we would need to create a
new one, which isn't worth the hassle seeing as it can't happen.