Add smoke test for config builder instrumentation issue (#6224)
## Summary of changes

Adds a smoke test for the issue fixed in #6147 

## Reason for change

The issue in #6147 is complex and subtle, with a high risk of
regression, so we want to have smoke tests to catch it`

## Implementation details

Add a sample app very similar to the one tested with in #6147. Without
the fix, [the test fails
with](https://dev.azure.com/datadoghq/dd-trace-dotnet/_build/results?buildId=166792&view=results)

```html
<b> Description: </b>An error occurred during the processing of a configuration file required to service this request. Please review the specific error details below and modify your configuration file appropriately.
<br><br>
<b> Parser Error Message: </b>The configBuilder 'CustomBuilder' failed while processing the configuration section 'appSettings'.: The ConfigurationBuilder 'CustomBuilder[Microsoft.Configuration.ConfigurationBuilders.CustomConfigBuilder]' has recursively re-entered processing of the 'appSettings' section.<br><br>
```

After the fix (now merged) the tests pass

## Test coverage

This is primarily a smoke test, so we confirm
- The sites are running and responding to HTTP requests correctly
- We instrument the application (it has logs)
- There are no errors in the logs

## Other details

For some reason, I had massive issues getting the app to "recognize" the
`DD_TRACE_AGENT_URL=http://test-agent.windows:8126` env var setting
that's required to talk to the test agent. I have no idea why, and the
only resolution I could find was to use powershell to set the variables
inside the docker image

---------

Co-authored-by: Kevin Gosse <kevin.gosse@datadoghq.com>