[ASM] Restrict RASP Lfi operations to read operation only (#6221)
## Summary of changes

In order to avoid too many calls to the WAF, after some discussions, it
was agreed that we should only call the WAF in RASP when accessing a
file for read operations.

## Reason for change

Performance.

## Implementation details

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->