[CI Visibility] GitInfoProvider refactor (#6259)
## Summary of changes

This PR is a complete refactor of the `GitInfo` class to a new structure
to allow multiple providers `IGitInfoProvider`. It also contains two
implementations:
1. GitInfoProvider based on calls to the git cli.
2. ManualParserGitInfoProvider based on manually parse the `.git` folder
without any git dependency (Without support for Delta objects).

## Reason for change

The previous structure was hard to maintain and to create fallbacks in
case a provider cannot read the git information. For example the
ManualParserGitInfoProvider has some issue in certain git folders (with
Delta objects), that's why we are using the one based on the git cli as
a fallback.

## Test coverage

Existing tests were modified and a new test was added.

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->