[ASM] Fix issue with waf disposing while it might concurrently be updating  (#6250)
## Summary of changes

When disposing the waf, sometimes an update might be taking place at the
same time, in a concurrent thread.
We shouldn't dispose while an update is taking place, so use the
ReaderWriterLock to dispose and block any access.

## Reason for change
Fix for this
[ticket](https://datadoghq.atlassian.net/jira/software/c/projects/APPSEC/boards/2876?selectedIssue=APPSEC-55734&atlOrigin=eyJpIjoiMmE0ZDM2ZjhhOTdjNGUyZGI5MDIwNmVkNjQ3MDlkNDkiLCJwIjoiaiJ9)
## Implementation details

## Test coverage
Added a unit test testing that update is either a success ,e ither fails
because dispose happened before.

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->