Add initial support for .NET 9 (#6265)
## Summary of changes

Updates as much as possible to support the .NET 9 RC2 release (with a
view to update to GA ASAP)

## Reason for change

.NET 9 is released very soon, and we want to support it

## Implementation details

As you may be able to tell, this was pretty painful, with a lot of small
cuts.
There are some things that I chose to skip rather than address.
Some of these are small... some of them are not 🙈 

It's probably best to review commits one at a time, so you can skip the
"boilerplate" ones

At a high level

- Updated all references to 8.0.100 SDK to the latest .NET 9 SDK ♻️ 
- Use different VM pools for safety 🦺 
- Add `net9.0` targets to all the apps that should support it ➕ 
- Fix a bunch of build warnings  🏗️ 
- Exclude new experimental activities added in .NET 9 so we don't get
duplicate spans 🛑
- Bump all the built-in integrations to use .NET 9 (introduced a shared
constant to make bumping that easier in the future) 🎉
- Update guard rails to allow injection into .NET 9 (could delay this a
little until we have ironed out the kinks) ❓
- Add support for new supported method overloads for IAST 🆕 
  - `File` has new overloads
  - `Type` has a new overload
- Skip unsupported method overloads for IAST tests 🦕  
- There are new overloads for `string` and `StringBuilder` that use
`ReadOnlySpan<>` - we don't currently support instrumenting those
- Additionally, due to the new overload selection in C#13, more methods
are selected that use `ReadOnlySpan<T>` instead of `string` which means
overall we have less coverage
- This is also annoying because it means the "wrong" methods are
selected in expressions, which causes compiler errors, so we have to
switch to explicitly passing arrays
- We should prioritise adding support for `ReadOnlySpan` for strings if
possible
- Skipped tests that can't be easily fixed (see below) ⚠️ 

## Test coverage

- Integration tests are all running on .NET 9
- _Some_ smoke tests (See details below)

## Other details

We have some outstanding issues to address ASAP after this is merged.
These are also marked with `FIXME: .NET 9`
- [ ] Our Linux crash tracking smoke tests hang on .NET 9 on x64
- [x] ~Our Linux arm64 smoke tests crash (I believe this is related to
running on docker compose v2, so I've updated the VMs)~
- Updating the VMs to use docker v2 fixed the issue. I will look to
update all our VMs to use v2 later, but that's non-blocking
- [ ] Debugger `ProbesTests` are failing in .NET 7+ - it looks like the
.NET 9 SDK generates slightly different code, this is probably only an
issue for the snapshots, so the snapshots should be updated and the test
unskipped
- [ ] .NET 9 changed how `WebRequest` works _significantly_. We will
need to update our integrations, but it's not clear how best to do that
without more exploration. This only affects the case where
`AllowWriteStreamBuffering=false` (which is not the default) so changed
the tests to use that for now instead.
- [ ] ManualInstrumentation on Windows .NET 9, x86 _only_ are failing
because they're failing to find the runtime. I think this is a niche
build issue that will be solved with the GA, so just skipped for now
- [ ] TrimmingTests on Linux debian, arm64 .NET 9 _only_ are crashing
while trying to add the startup hook. Struggling to get any useful logs
or crash dumps right now, so skipping and will look into it closer
(potentially on mac arm64 where easier to test)
- [ ] Profiler execution tests are currently broken - they need to have
the .NET 9 SDK installed, we could probably safely wait for the GA
release seeing as that's only 2 days away...
- [ ] Gitlab build will be broken... I need to follow up and fix that
ASAP now