[ASM] Serialize ASM tags and metrics (#6272)
## Summary of changes

Some of the ASM tags and metrics that are added to the span can be added
during the span's serialization. This change would increase the
performance of such operations.

## Reason for change

## Implementation details

## Test coverage

The integration tests already test the presence of such tags and
metrics, so passing them with this approach would validate the
implementation.

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->