Add support for `AllowWriteStreamBuffering=false` in `WebRequest` for .NET 9 (#6271)
## Summary of changes

Add support for `AllowWriteStreamBuffering=false` in `WebRequest` in
.NET 9

## Reason for change

Prior to .NET 9, the `AllowWriteStreamBuffering` flag did nothing in
.NET Core. [Support was implemented for .NET
9](https://github.com/dotnet/runtime/pull/95001/files), but
unfortunately this has a significant adverse affect on our
instrumentation.

_In general_ the `HttpWebRequest` support in .NET Core is kinda hacky.
It's essentially a wrapper over the top of an underlying `HttpClient`.
Unfortunately, many of the APIs it provides are problematic for us,
because they require calling two separate methods, e.g.
`GetRequestStream()` followed by `GetResponseStream()`. The definition
of exactly where the Span operations should go becomes kinda fuzzy, and
we end up having to use `ConditionalWeakTables` and using the web
request itself as a carrier to flow context around.

Up to now, that mostly worked because the
`GetRequestStream()`/`BeginGetRequestStream()` methods didn't actually
make an HTTP request, it just "prepared" the request, and the request
was only sent when you call `GetResponse`. However, with
`AllowWriteStreamBuffering=false` that's no longer the case; now
`GetRequestStream()` _does_ send a request. The upshot is that without
this PR we end up with _two_ HTTP spans...

## Implementation details

To work around the issue (and I consider it a workaround, we should
maybe reconsider the implementation entirely for .NET 9), I created
three new integrations which are specific to .NET 9+

- `BeginGetRequestStream()`
- `GetRequestStream()`
- `EndGetResponse()`

In general, these work the same as the .NET 8 versions, but with some
additional behaviour when `AllowWriteStreamBuffering=false`:

- When `AllowWriteStreamBuffering=false`, `BeginGetRequestStream()` and
`GetRequestStream()` inject the context headers into the request as
usual, but then _also_ briefly set `x-datadog-tracing-enabled:false` in
`OnMethodBegin`
- Setting the don't trace flag ensures the `HttpClient` doesn't create a
"real" span, so we don't get duplicate spans. Note that we can't easily
use the `HttpClient` as "the" span, because we don't read the response
etc until a _different_ method call, which risks leaking spans etc.
- Note that the propagation headers _are_ still injected, so propagation
should still work as normal, we're just disabling creating a span for
it.
- In `OnMethodEnd` we remove the `x-datadog-tracing-enabled:false`
header again. It's a bit weird to modify the request headers again
_after_ the request has been sent, but `WebRequest` copies the headers
into the `HttpClient` collection anyway, so it doesn't really matter
AFAICT.
- In <.NET 9, `BeginGetResponse()`/`EndGetResponse()` relies on the fact
that the _actual_ underlying `HttpClient` call is made here. However, in
.NET 9 that's not the case when `AllowWriteStreamBuffering=false`.
Without changes, this span ends up missing.
- To work around this, added a .NET 9+ instrumentation of
`EndGetResponse()` (previously this instrumentation was .NET FX only).
- The instrumentation is largely identical to the .NET FX
instrumentation
- The one key difference is that in .NET FX we can get when the request
was started (to calculate the span duration). That's not available in
.NET Core 🙁
- To work around that, added an additional header with the creation
timestamp to the `WebHeaderCollection` in
`BeginGetRequestStream.OnMethodEnd` and `GetRequestStream.OnMethodEnd`
- These are _only_ added when `AllowWriteStreamBuffering=false`, so they
are only added _after_ the initial request is sent
  - That means the _won't_ be propagated. 
- The main risk here ACAIT is that someone "messes" with the headers in
between calling `BeginGetRequestStream()` (which sends the request) and
`EndGetResponse()`. But worst case, if they do, they get incorrect
timings. Plus it doesn't really make _sense_ to mess with them

## Test coverage

Added additional test case to the `WebRequestTests` for testing
`AllowWriteStreamBuffering=false` and updated the snapshots. These work
as expected in < .NET 9.

I had to do one small hack for the .NET 9 snapshots. Prior to .NET 9,
when you use `BeginRequestResponse()` and `EndGetResponse()`, the span
we create is created by the underlying `HttpClient`. In .NET 9, we
create it "manually" in `WebRequest` instead of letting the underlying
`HttpClient` handle it. Functionally this is fine, but it messes with
the snapshots slightly

Rather than create a whole new set of snapshots for .NET 9 (which makes
it harder to see differences), I just manually tweaked the span in the
test before running snapshots (changing the `component` and
`http-client-handler-type` tags). I think this is the best option
personally

## Other details

Did some minor refactoring into `WebRequestCommon` too. Each commit is
atomic as usual

Stacked on 
- https://github.com/DataDog/dd-trace-dotnet/pull/6265