[CI Visibility] Fix GitCommandParser test (#6289)
## Summary of changes

The PR fixes the GitCommand Parser test

## Reason for change

The test is failing on some platforms because we work with a detached
Head, and that makes the Branch to be null.

## Implementation details

Removes the branch assertion.

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->