Ensure that we never run any call target instrumentations in partial trust (#6290)
## Summary of changes

Update the `CallTargetInvoker` to explicitly bail-out if partial-trust
is detected.

## Reason for change

We don't support partial trust. In the native profiler, we have code
that explicitly bails out if we're in partial trust. Unfortunately...
IIS.

In IIS, we could have two apps running in the same app pool&mdash;one in
full trust, one in partial trust. The full trust app will be
instrumented as normal. Unfortunately, because they're running in the
same app pool, _the partial trust app is effectively also instrumented_.

In this scenario, we'll likely create errors in the partial trust app
(we have seen `MemberAccessException`s for example), and we still can't
_really_ work there.

## Implementation details

The problem is essentially the same as a recent fixed issue (#6147) -
multiple apps in the same pool causing oddities with instrumentation.
The fix in this PR piggy-backs in the work there, essentially

- We check whether the app domain is in partial trust.
- If it is, we bail out of running any CallTarget instrumentation.

This should solve the issue for calltarget and for manual
instrumentation (which is just calltarget in v3).

The issue remains for callsite instrumentation unfortunately, and likely
cannot easily be mitigated @daniel-romano-DD

## Test coverage

_sigh_ I spent all day trying to create an aspnet app that I could run
in partial trust. All day. I failed. If anyone can produce one that will
run in IIS under partial trust, then I have the setup prepared.

Unfortunately, as I can't even pass that low bar, I can't even
_manually_ test this 🙁 All I can confirm is that it doesn't affect "full
trust" applications that we use in CI 😅

## Other details

Seriously, can _anyone_ get a partial trust app running? 🤷‍♂️