[CI Visibility] Add support for coverlet.msbuild coverage reporting (#6284)
## Summary of changes

This PR add support to automatic reporting code coverage percentage from
the `coverlet.msbuild` package.

## Reason for change

We added support to automatic reporting from the coverlet datacollector
but we were missing the automatic reporting from the coverlet msbuild
task.

## Implementation details

For some reason we are hitting an edge case in the DuckType library were
`Type.GetType` returns null when using a partial type name and not a
qualified one. I did a modification to the ducktype library to handle
this edge case.

## Test coverage

Works on my machine 🤷🏻  🤣 .

Now seriously... creating a test for this specific feature will require
a refactor to the way we run/publish the samples apps. Currently we
build a publish them in a different step to the artifacts folder and run
them later in the test. `coverlet.msbuild` is not compatible with that
because it contains a series of msbuild tasks that applies new targets
to build the covered binary and then get the coverage values after the
execution. Something that we cannot do right now.

Because this is a niche, best effort, and easy to test locally
functionality. I decided to skip a custom test for this one. I'm
evaluating on adding the test in the Test Visibility test-environment
that it would be easier to do.

For the ducktyping change if this PR passes we are ok to go because a
change of behaviour in the normal case is not expected.

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->