[CI Visibility] Catch `UploadRepositoryChanges` exceptions (#6331)
## Summary of changes

This PR adds some missing try/catch for the UploadRepositoryChanges
calls.

## Reason for change

This method is being used in the `dd-trace` runner, currently if the
method fails (backend error) then `dd-trace` will throw an exception.

## Implementation details

We create a function method to wrap the call with a try/catch block so
we don't throw and fail the customer CI.

## Test coverage

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->