[Crashtracking] Filter the ptrace_create hook (#6337)
## Summary of changes

When marking a crash as suspicious or not, ignore the entrypoint
inserted by the ptrace_create hook.

## Reason for change

The entrypoint is inserted into every new thread. Because of that,
pretty much every crash was marked as suspicious.

## Implementation details

Because the crashtracker doesn't have access to symbols, we had to
publicly expose the inserted entrypoint. To filter it more easily, it
has been renamed to `dd_pthread_entry`.

## Test coverage

Added a new test case that triggers a crash in a new thread. The issue
wasn't discovered because the existing test was crashing in the main
thread, so it didn't have the injected entrypoint.

---------

Co-authored-by: Gregory LEOCADIE <gregory.leocadie@datadoghq.com>