Fix bug in manual instrumentation (#6330)
## Summary of changes

Fixes the `Tracer.Settings` getter in manual instrumentation

## Reason for change

While investigating something else, I noticed that the `Tracer.Settings`
getter doesn't work correctly if the tracer has been reconfigured. The
previous implementation assumed that the automatic `Tracer` instance is
tied to a given `ImmutableTracerSettings` object, but that's not the
case. It's the `TracerManager` that's tied to a settings object, but in
"normal" usage, the `Tracer` instance _isn't_ tied to a single
`TracerManager`, and hence _isn't_ tied to a settings object. e.g

```csharp
Tracer.Instance.ActiveScope;// Trigger the client-side creation of the `Tracer` object with default settings

// customer manually re-configures tracer
var settings = TracerSettings.FromDefaultSources();
settings.ServiceName = "updated-name";
settings.Environment = "updated-env";
settings.GlobalTags = new Dictionary<string, string> { { "Updated-key", "Updated Value" } };
Tracer.Configure(settings); // update Tracer.Instance

// Fetch the settings back
var immutable = Tracer.Instance.Settings;
Debug.Assert(immutable.ServiceName == settings.ServiceName); // fails
Debug.Assert(immutable.Environment == settings.Environment ); // fails
// The settings have not changed from the _original_ values
```

## Implementation details

Instead of populating a settings object when creating the manual
`Tracer` instance, create it when the customer explicitly calls
`Tracer.Settings`. This is likely to be rare (will add telemetry in a
followup PR) but we also don't want to create the dictionary transport
every time they call it. To work around this, we track the "current"
(automatic) `ImmutableTracerSettings` in a local field and pass that to
the automatic side, so that it can check whether the settings need to be
updated or not.

## Test coverage

Added additional assertions to the manual instrumentation tests, which
fail prior to these changes

## Other details

This _removes_ an instrumentation point (and adds another) but we keep
the instrumentation around for back-compat reasons