Add support for automatically detecting OpenTelemetry Instrumentation Libraries, and use it to initialize OpenTelemetry.Instrumentation.Hangfire v1.6.0-beta.1+
This support is demonstrated through passing the regression test after removing the manual  ".AddHangfireInstrumentation()" code that was previously needed in the instrumented application. Even without the OpenTelemetry SDK, we are able to see traces emitted from the instrumentation library.

Note: Changes in the snapshots are due to missing tags that the OpenTelemetry SDK was previously emitting.