[Profiler] Update tests to save output to xunit.txt files (#6160)
## Summary of changes
Update all tests to send output to console AND file

## Reason for change
That way, it is possible to have the tests output visible in the CI with
xunit.txt file to download

## Implementation details
Update all tests

## Test coverage
n/a

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->