Fix alpine .NET Core 3.1 telemetry failures (#6339)
## Summary of changes

Fix the telemetry issues we saw after updating the VMs recently

## Reason for change

We updated the VMs and suddenly .NET Core <=3.1 started failing on
alpine when we used the "standalone telemetry agent", only when using
the "curl" handler (not the socket handler).

We eventually traced it down to the fact we were setting the url to
`localhost`, not `127.0.0.1` as we do elsewhere. Presumably an update to
the underlying libcurl caused it to choose the IPv6 address, which the
listener is not bound to, causing telemetry to fail

## Implementation details

- Reverted the change in
https://github.com/DataDog/dd-trace-dotnet/pull/6324
- `localhost` -> `127.0.0.1` in the `TelemetryHelper`

## Test coverage

The same

## Other details

Reverting the change isn't necessary obviously, but it shows that we
fixed it.