Try to fix flake in ManualInstrumentation tests (#6333)
## Summary of changes

Add a mutex to try to avoid flake

## Reason for change

We see occasional flake in the manual + automatic instrumentation tests.
I don't entirely understand why this happens _other than_ it shows that
the httplistener span is getting serialized _after_ we have reconfigured
the tracer (because traces _aren't_ tied to a settings instance). This
is arguably a bug, but it's "known" to an extent. The problem is that
it's causing flake when you change the base service name.

> FWIW, we already "fixed" this same symptom once:
https://github.com/DataDog/dd-trace-dotnet/pull/4960


## Implementation details

Add a mutex to try to make sure the httplistener span is closed before
we close the request span

## Test coverage

Enabled debug logs so that we can get more details if this _does_ recur.

## Other details

Stacked on:
- https://github.com/DataDog/dd-trace-dotnet/pull/6330