[CI Visibility] Specify if the user is setting the DD_SERVICE (#6348)
## Summary of changes

This PR sets a tag to let the ci visibility backend know if the
DD_SERVICE value was provided by the user or auto-generated.

## Reason for change

This required by the backend.

## Test coverage

- Updated snapshots.
- Fixed tests

## Other details
<!-- Fixes #{issue} -->

<!-- ⚠️ Note: where possible, please obtain 2 approvals prior to
merging. Unless CODEOWNERS specifies otherwise, for external teams it is
typically best to have one review from a team member, and one review
from apm-dotnet. Trivial changes do not require 2 reviews. -->