[ASM] iast: Tainting of DefaultInterpolatedStringHandler (#6340)
## Summary of changes

This PR introduce the support of `DefaultInterpolatedStringHandler` for
IAST.
The resulting strings of `DefaultInterpolatedStringHandler` will now be
tainted.

## Reason for change

Since the release of .NET Core 6, interpolated strings got a performance
optimisation and now use
[DefaultInterpolatedStringHandler](https://docs.microsoft.com/en-us/dotnet/api/system.runtime.compilerservices.defaultinterpolatedstringhandler?view=net-6.0)
to build strings.

Some vulnerabilities couldn't be detected because strings built with
`DefaultInterpolatedStringHandler` weren't tainted.

## Implementation details

As `DefaultInterpolatedStringHandler` is a ref struct, we call some IL
to get it's own stack pointer value and taint it.
We need to get that pointer value to track it and its tainted sources.

## Test coverage

- New unit tests were added:
- testing all aspects with an explicit call to the
`DefaultInterpolatedStringHandler` ref struct
- testing implicit interpolated strings (`$""`) in various complex cases
- Aspects tests added for `IastInstrumentationUnitTests`

## Other details

This PR don't handle the correct values for `start` and `length` of
tainted Sources.

---------

Co-authored-by: Daniel Romano <108014683+daniel-romano-DD@users.noreply.github.com>