Bail out if we're in a no-dynamic-code scenario (#6362)
## Summary of changes

Bail out in the managed loader if we're in a no-dynamic code scenario

## Reason for change

Our automatic instrumentation is currently inherently tied to emitting
dynamic code (Duck typing). In no-dynamic-code scenarios we will likely
end up crashing (or giving a very degraded experience, with a
performance hit at the very least). It's better to bail out early than
do that

## Implementation details

Check for the
`System.Runtime.CompilerServices.RuntimeFeature.IsDynamicCodeSupported`
app context switch in the managed loader. This controls the
`RuntimeFeature.IsDynamicCodeSupported` flag, which in turn determines
whether the app auto-bails out in .NET 8+.

Note that this flag is set when you add `<PublishAot>` to your project
_even if you don't publish with AOT_. I imagine there are some other
environments (pre .NET 8) in which dynamic code is not available, but I
don't know if we need to worry about them

## Test coverage

Added an integration tests that sets the flag and confirms we don't get
traces when the flag is set to `false`

## Other details

I had to add a helper to the `TestHelper` to pass runtime arguments to
`dotnet`